requirements.txt
Django==4.2.7
#mysqlclient==2.2.0
python-decouple==3.8